Added separate class for setup and teardown that is run before any test in stepdef. Populated more Trainee tests and added method to implementation to abstract code from stepdefs.
