--- src/test/resources/features/trainee/userStory-57.feature
@@ -1,7 +1,0 @@
-Scenario: Adding comments as a trainee
-
-Given I am logged in as a trainee
-
-When I select a week
-
-Then I am able to add a comment
