--- src/test/resources/features/trainee/userStory-81.feature
@@ -1,7 +1,0 @@
-Scenario: The Trainee wants to see if their feedback from is in progress or late
-
-Given The trainee is logged in successfully
-
-When The trainee clicks on the home page
-
-Then The trainee will be able to view the status of their feedback form
